% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{calculateFoldChange}
\alias{calculateFoldChange}
\title{Determine fold changes and map to finite numbers.}
\usage{
calculateFoldChange(abundances1, abundances2)
}
\arguments{
\item{abundances1}{first abundance vector}

\item{abundances2}{second abundance vector}
}
\value{
fold change
}
\description{
fc = log2(abundances1/abundances2)
Should be NaN-save, i.e. NaN in any of the two abundance vectors results in NaN in the fold change vector.
NaN = not quantifiable, hence fc = NaN = not quantifiable
}
