% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{checkAccessionFormat}
\alias{checkAccessionFormat}
\title{Check that the protein accession is of the format ..|..|..}
\usage{
checkAccessionFormat(accession)
}
\arguments{
\item{accession}{protein accession}
}
\value{
Boolean. Is the accession of the format ..|..|.. ?
}
\description{
Note that NA returns TRUE.
}
