% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{countOccurrences}
\alias{countOccurrences}
\title{Count the occurences of character char in string s.}
\usage{
countOccurrences(char, s)
}
\arguments{
\item{char}{character}

\item{s}{string}
}
\value{
number of occurences
}
\description{
Count the occurences of character char in string s.
}
