% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{cutSequence}
\alias{cutSequence}
\title{Reduces the length of the peptide sequence to the first 25 amino acids.}
\usage{
cutSequence(s)
}
\arguments{
\item{s}{peptide sequence}
}
\value{
shorter peptide sequence
}
\description{
Reduces the length of the peptide sequence to the first 25 amino acids.
}
