% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{findPeptidesOfInterest}
\alias{findPeptidesOfInterest}
\title{Find all peptides of interest.}
\usage{
findPeptidesOfInterest(data)
}
\arguments{
\item{data}{peptide dataframe}
}
\value{
dataframe containing only peptides of interest.
}
\description{
We assume a <peptides.of.interest> vector exists.
}
