% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{findProteinsOfInterest}
\alias{findProteinsOfInterest}
\title{Find all proteins of interest.}
\usage{
findProteinsOfInterest(data)
}
\arguments{
\item{data}{peptide dataframe}
}
\value{
dataframe containing only peptides which occur in <proteins.of.interst>
}
\description{
We assume a <proteins.of.interest> vector exists.
}
