% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{getAccession}
\alias{getAccession}
\title{Extracts the second entry from a string of the form ..|..|..}
\usage{
getAccession(string)
}
\arguments{
\item{string}{string of the form x|y|z}
}
\value{
y
}
\description{
Extracts the second entry from a string of the form ..|..|..
}
