% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{getAverageIntensity}
\alias{getAverageIntensity}
\title{Returns an average peptide intensity over all study variables.}
\usage{
getAverageIntensity(data)
}
\arguments{
\item{data}{dataframe with columns "peptide_abundance_study_variable[*]"}
}
\value{
mean intensity over all channels
}
\description{
Returns an average peptide intensity over all study variables.
}
