% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{getGene}
\alias{getGene}
\title{Extracts the third entry from a string of the form ..|..|..}
\usage{
getGene(string)
}
\arguments{
\item{string}{string of the form x|y|z}
}
\value{
z
}
\description{
Extracts the third entry from a string of the form ..|..|..
}
