% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{getModsSummary}
\alias{getModsSummary}
\title{Create a summary table of all modifications and their specificities.}
\usage{
getModsSummary(data)
}
\arguments{
\item{data}{peptide dataframe}
}
\value{
summary table
}
\description{
Required input is a dataframe with a "sequence" and "modifications" column in mzTab standard.
}
