% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{getPCA}
\alias{getPCA}
\title{Calculate the principal component object.}
\usage{
getPCA(data)
}
\arguments{
\item{data}{peptide dataframe}
}
\value{
principal component object
}
\description{
Calculate the principal component object.
}
