% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{getPCAeigenvector}
\alias{getPCAeigenvector}
\title{Eigenvectors point in the direction of the principal componets in the high-dimensional peptide abundance space.}
\usage{
getPCAeigenvector(pca, n)
}
\arguments{
\item{pca}{principal component object, see getPCA()}

\item{n}{number of principal component}
}
\value{
row indices of peptides in original peptide dataframe, see getPCA()
}
\description{
Important peptides (i.e. the ones with a large absolute eigenvector component) contribute most to this principal component.
The function returns these peptides i.e. their row index.
}
