% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{getPeptideQuants}
\alias{getPeptideQuants}
\title{Returns a dataframe containing only the peptide quantification columns.}
\usage{
getPeptideQuants(data)
}
\arguments{
\item{data}{dataframe with columns "peptide_abundance_study_variable[*]"}
}
\value{
quants only dataframe
}
\description{
Returns a dataframe containing only the peptide quantification columns.
}
