% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{getQuantSummary}
\alias{getQuantSummary}
\title{Create a summary table of all quantifications.}
\usage{
getQuantSummary(data)
}
\arguments{
\item{data}{peptide dataframe}
}
\value{
summary table
}
\description{
How many quantifications are finite, zero or NaN in each sample?
}
