% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{isEmpty}
\alias{isEmpty}
\title{Check if the vector/column is empty.}
\usage{
isEmpty(column)
}
\arguments{
\item{column}{column to be checked}
}
\value{
Boolean. Is the entire column empty?
}
\description{
i.e. all entries are NA or "" etc. or the the vector is of length 0
}
