% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{makeModifiedSequenceChargeUnique}
\alias{makeModifiedSequenceChargeUnique}
\title{Makes the (modified sequence, charge) combination unique by picking the quants with maximum intensity.}
\usage{
makeModifiedSequenceChargeUnique(data)
}
\arguments{
\item{data}{peptide dataframe}
}
\value{
subset of input dataframe with double rows removed
}
\description{
Makes the (modified sequence, charge) combination unique by picking the quants with maximum intensity.
}
