% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{numberOfStudyVariables}
\alias{numberOfStudyVariables}
\title{Returns the number of quantification channels i.e. the number of "peptide_abundance_study_variable[*]" columns.}
\usage{
numberOfStudyVariables(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
number of quantification channels
}
\description{
Returns the number of quantification channels i.e. the number of "peptide_abundance_study_variable[*]" columns.
}
