% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Plot boxplot of all peptide quantifications.}
\usage{
plotBoxplot(data, pdf.file)
}
\arguments{
\item{data}{peptide dataframe}

\item{pdf.file}{path to output pdf file}
}
\description{
Plot boxplot of all peptide quantifications.
}
