% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotChargeDistribution}
\alias{plotChargeDistribution}
\title{Plot charge distribution.}
\usage{
plotChargeDistribution(data, pdf.file)
}
\arguments{
\item{data}{peptide dataframe}

\item{pdf.file}{path to output pdf file}
}
\description{
We assume a 'charge' column exists and do not check.
}
