% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotCorrelations}
\alias{plotCorrelations}
\title{Plot correlation of all peptide quantifications.}
\usage{
plotCorrelations(data, pdf.file)
}
\arguments{
\item{data}{peptide dataframe}

\item{pdf.file}{path to output pdf file}
}
\value{
correlation
}
\description{
Plot correlation of all peptide quantifications.
}
