% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotDeltaFcLogIntensity}
\alias{plotDeltaFcLogIntensity}
\title{Plot the difference between 2+ and 3+ fold changes of the same (modified) sequence against the logarithm of the average peptide intensity.}
\usage{
plotDeltaFcLogIntensity(data, sample.1, sample.2, pdf.file)
}
\arguments{
\item{data}{data frame}

\item{sample.1}{number of the sample i.e. study variable}

\item{sample.2}{number of the sample i.e. study variable}

\item{pdf.file}{path to output pdf file}
}
\description{
Plot the difference between 2+ and 3+ fold changes of the same (modified) sequence against the logarithm of the average peptide intensity.
}
