% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot distribution.}
\usage{
plotDistribution(vector, label, pdf.file)
}
\arguments{
\item{vector}{quantity for distribution}

\item{label}{label for x-axis}

\item{pdf.file}{path to output pdf file}
}
\description{
Plot distribution.
}
