% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotElutionTimeDistribution}
\alias{plotElutionTimeDistribution}
\title{Plot the distribution of peptide elution times.}
\usage{
plotElutionTimeDistribution(data, pdf.file)
}
\arguments{
\item{data}{dataframe with retention_time_window column}

\item{pdf.file}{path to output pdf file}
}
\description{
Each peptide reports a minimum/maximum retention time in the retention_time_window column.
}
