% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotFcLogIntensity}
\alias{plotFcLogIntensity}
\title{Plot fold change vs log intensity.}
\usage{
plotFcLogIntensity(fc.vector, intensity.vector, fc.label, pdf.file)
}
\arguments{
\item{fc.vector}{fold change vector (x-axis)}

\item{intensity.vector}{intensity vector (y-axis)}

\item{fc.label}{label for x-axis}

\item{pdf.file}{path to output pdf file}
}
\description{
Plot fold change vs log intensity.
}
