% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotFcLogIntensitySingleProtein}
\alias{plotFcLogIntensitySingleProtein}
\title{Plot peptide fold change vs log intensity for a single specific protein.}
\usage{
plotFcLogIntensitySingleProtein(data, protein, sample.1, sample.2,
  pdf.file)
}
\arguments{
\item{data}{data frame}

\item{protein}{protein}

\item{sample.1}{number of the sample i.e. study variable}

\item{sample.2}{number of the sample i.e. study variable}

\item{pdf.file}{path to output pdf file}
}
\description{
(same peptide sequence -> same colour)
}
