% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotFcSingleProtein}
\alias{plotFcSingleProtein}
\title{Plot sample (or group) index vs fold change for all peptides of a specific protein.}
\usage{
plotFcSingleProtein(data, protein, pdf.file)
}
\arguments{
\item{data}{data frame}

\item{protein}{protein}

\item{pdf.file}{path to output pdf file}
}
\description{
The fold change is calculated relative to the sample with the most peptide quantifications.
}
