% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotFrequencyOfFrequencies}
\alias{plotFrequencyOfFrequencies}
\title{Plot frequency of frequencies of any vector. Take for example a vector of protein accessions.}
\usage{
plotFrequencyOfFrequencies(vector, pdf.file, xlab = "frequency",
  ylab = "frequency of frequency", log = "y")
}
\arguments{
\item{vector}{any set, such as a vector of protein accessions}

\item{pdf.file}{path to output pdf file}
}
\description{
frequency: How often does a particular protein X occur?
frequency of frequencies: How often does a protein occur twice or three times and so on?
}
