% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotKendrick}
\alias{plotKendrick}
\title{Plot Kendrick nominal fractional mass plot.}
\usage{
plotKendrick(mass, pdf.file)
}
\arguments{
\item{mass}{peptide masses}

\item{pdf.file}{path to output pdf file}
}
\description{
Plot Kendrick nominal fractional mass plot.
}
