% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotPCAcomponents}
\alias{plotPCAcomponents}
\title{Plot the standard deviation of all principal components.}
\usage{
plotPCAcomponents(pca, pdf.file)
}
\arguments{
\item{pca}{principal component object, see getPCA()}

\item{pdf.file}{path to output pdf file}
}
\description{
Plot the standard deviation of all principal components.
}
