% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotPCAeigenvector}
\alias{plotPCAeigenvector}
\title{Plot the coordinates of the nth eigenvector.}
\usage{
plotPCAeigenvector(pca, data, n, pdf.file)
}
\arguments{
\item{pca}{principal component object, see getPCA()}

\item{data}{peptide dataframe}

\item{n}{number of principal component}

\item{pdf.file}{path to output pdf file}
}
\description{
Plot the coordinates of the nth eigenvector.
}
