% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotPCAscatter}
\alias{plotPCAscatter}
\title{Plot the scatter plot of the first n.pca principal components.}
\usage{
plotPCAscatter(pca, pdf.file)
}
\arguments{
\item{pca}{principal component object, see getPCA()}

\item{pdf.file}{path to output pdf file}
}
\description{
Plot the scatter plot of the first n.pca principal components.
}
