% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotPeptidesOfInterest}
\alias{plotPeptidesOfInterest}
\title{Plots the reported peptide abundances of all peptides of interest.}
\usage{
plotPeptidesOfInterest(data, pdf.file)
}
\arguments{
\item{data}{peptide dataframe}

\item{pdf.file}{path to output pdf file}
}
\description{
Plots the reported peptide abundances of all peptides of interest.
}
