% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotPeptidesPerProtein}
\alias{plotPeptidesPerProtein}
\title{Plot quantified peptides per protein vs frequency.}
\usage{
plotPeptidesPerProtein(data, pdf.file)
}
\arguments{
\item{data}{peptide dataframe}

\item{pdf.file}{path to output pdf file}
}
\description{
Plot quantified peptides per protein vs frequency.
}
