% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotProteinsOfInterest}
\alias{plotProteinsOfInterest}
\title{Plots the reported peptide abundances of all proteins of interest.}
\usage{
plotProteinsOfInterest(data, pdf.file)
}
\arguments{
\item{data}{peptide dataframe}

\item{pdf.file}{path to output pdf file}
}
\description{
Plots the reported peptide abundances of all proteins of interest.
}
