% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{plotQuantFrequency}
\alias{plotQuantFrequency}
\title{Plot (in)complete quantifications.}
\usage{
plotQuantFrequency(quants, pdf.file)
}
\arguments{
\item{quants}{dataframe with peptide quantifications, see getPeptideQuants()}

\item{pdf.file}{path to output pdf file}
}
\description{
Not all peptides need to be quantified in all channels/samples. See for example knock-out or TAILS experiments.
Not quantified can mean either NaN or exactly zero.
The plot below summarises how many peptides were quantified in x samples. 1 <= x <= number of samples
}
