% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{readMzTabPEP}
\alias{readMzTabPEP}
\title{Read the PEP section of an mzTab file.}
\usage{
readMzTabPEP(file)
}
\arguments{
\item{file}{path to mzTab file}
}
\value{
dataframe of PEP section
}
\description{
Read the PEP section of an mzTab file.
}
