% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{splitAccession}
\alias{splitAccession}
\title{Splits fasta protein accession into UniProt accession and gene name.}
\usage{
splitAccession(peptide.data)
}
\arguments{
\item{peptide.data}{dataframe with <accession> column}
}
\value{
dataframe with UniProt accession and gene name
}
\description{
Splits fasta protein accession into UniProt accession and gene name.
}
