% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{startSection}
\alias{startSection}
\title{Simple Moving Average without leading NA.}
\usage{
startSection(file, section.identifier)
}
\arguments{
\item{file}{path to mzTab file}

\item{section.identifier}{identifier at the start of the section, either 'PEH', 'PRH' or 'PSH'}
}
\value{
first row of the section
}
\description{
Simple Moving Average without leading NA.
}
