% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{studyVariableExists}
\alias{studyVariableExists}
\title{Check if a specific "peptide_abundance_study_variable[n]" column exists.}
\usage{
studyVariableExists(data, n)
}
\arguments{
\item{data}{dataframe}

\item{n}{index of study variable}
}
\value{
Boolean. Does the column "peptide_abundance_study_variable[n]" exist?
}
\description{
Check if a specific "peptide_abundance_study_variable[n]" column exists.
}
