% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabHelper.R
\name{uniqueColors}
\alias{uniqueColors}
\title{Returns a unique colour for each string.}
\usage{
uniqueColors(string.vector)
}
\arguments{
\item{string.vector}{vector of strings}
}
\value{
colours
}
\description{
Returns a unique colour for each string.
}
